fix: update import paths after moving continuedev into ghost folder
