+++ src/api/providers/__tests__/mistral-fim.spec.ts
@@ -6,10 +6,10 @@ vitest.mock("vscode", () => ({}))
 
 import { MistralHandler } from "../mistral"
 import { ApiHandlerOptions } from "../../../shared/api"
-import { streamSse } from "../../../services/continuedev/core/fetch/stream"
+import { streamSse } from "../../../services/ghost/continuedev/core/fetch/stream"
 
 // Mock the stream module
-vitest.mock("../../../services/continuedev/core/fetch/stream", () => ({
+vitest.mock("../../../services/ghost/continuedev/core/fetch/stream", () => ({
 	streamSse: vitest.fn(),
 }))
