+++ src/services/ghost/GhostJetbrainsBridge.ts
@@ -4,7 +4,7 @@ import { z } from "zod"
 import { GhostServiceManager } from "./GhostServiceManager"
 import { ClineProvider } from "../../core/webview/ClineProvider"
 import { getKiloCodeWrapperProperties } from "../../core/kilocode/wrapper"
-import { languageForFilepath } from "../continuedev/core/autocomplete/constants/AutocompleteLanguageInfo"
+import { languageForFilepath } from "./continuedev/core/autocomplete/constants/AutocompleteLanguageInfo"
 import { GhostContextProvider } from "./types"
 import { FimPromptBuilder } from "./classic-auto-complete/FillInTheMiddle"
 import { HoleFiller } from "./classic-auto-complete/HoleFiller"
