+++ src/services/ghost/classic-auto-complete/HoleFiller.ts
@@ -6,7 +6,7 @@ import {
 	ChatCompletionResult,
 } from "../types"
 import { getProcessedSnippets } from "./getProcessedSnippets"
-import { formatSnippets } from "../../continuedev/core/autocomplete/templating/formatting"
+import { formatSnippets } from "../continuedev/core/autocomplete/templating/formatting"
 import { GhostModel } from "../GhostModel"
 import { ApiStreamChunk } from "../../../api/transform/stream"
